feat: entity delete scheduler 통합
